feat: moved new gaussian policy outside of experimental package
